+++ next/src/components/templates/TemplateData.tsx
@@ -1,0 +1,67 @@
+export const TEMPLATE_DATA: TemplateModel[] = [
+    { 
+        name: "PlatformerGPT  🎮", 
+        value: "Write some code to make a platformer game.", 
+    },
+    { 
+        name: "ResearchGPT  📜", 
+        value: "Create a comprehensive report of a topic of your choice", 
+    },
+    { 
+        name: "TravelGPT  🌴", 
+        value: "Plan a detailed trip to destination", 
+    },
+    { 
+        name: "PlatformerGPT  🎮", 
+        value: "Write some code to make a platformer game.", 
+    },
+    { 
+        name: "ResearchGPT  📜", 
+        value: "Create a comprehensive report of a topic of your choice", 
+    },
+    { 
+        name: "TravelGPT  🌴", 
+        value: "Plan a detailed trip to destination", 
+    },
+    { 
+        name: "PlatformerGPT  🎮", 
+        value: "Write some code to make a platformer game.", 
+    },
+    { 
+        name: "ResearchGPT  📜", 
+        value: "Create a comprehensive report of a topic of your choice", 
+    },
+    { 
+        name: "TravelGPT  🌴", 
+        value: "Plan a detailed trip to destination", 
+    },
+    { 
+        name: "PlatformerGPT  🎮", 
+        value: "Write some code to make a platformer game.", 
+    },
+    { 
+        name: "ResearchGPT  📜", 
+        value: "Create a comprehensive report of a topic of your choice", 
+    },
+    { 
+        name: "TravelGPT  🌴", 
+        value: "Plan a detailed trip to destination", 
+    },
+    { 
+        name: "PlatformerGPT  🎮", 
+        value: "Write some code to make a platformer game.", 
+    },
+    { 
+        name: "ResearchGPT  📜", 
+        value: "Create a comprehensive report of a topic of your choice", 
+    },
+    { 
+        name: "TravelGPT  🌴", 
+        value: "Plan a detailed trip to destination", 
+    },
+];
+
+export interface TemplateModel {
+    name: string;
+    value: string;
+}
