+++ next/src/pages/templates.tsx
@@ -6,60 +6,12 @@ import nextI18NextConfig from "../../next-i18next.config.js";
 import { languages } from "../utils/languages";
 import { GetStaticProps } from "next";
 import AppTitle from "../components/AppTitle";
+import { TEMPLATE_DATA } from "../components/templates/TemplateData";
+import TemplateCard from "../components/templates/TemplateCard";
+import FadeIn from "../components/motions/FadeIn";
 
-const categoryColors = {
-    "Games": "bg-teal-400",
-    "Academics": "bg-pink-900",
-    "Travel": "bg-blue-500",
-    "Web": "bg-green-600",
-    "Social Media": "bg-yellow-500",
-    "Communications": "bg-indigo-500",
-    "Career": "bg-purple-700",
-    "Creative Writing": "bg-red-500",
-    "Health & Wellness": "bg-green-500",
-    "Business": "bg-blue-600",
-    "Personal Finance": "bg-yellow-600",
-    "News & Media": "bg-red-600",
-    "Event Planning": "bg-indigo-600",
-    "Blogging": "bg-purple-500"
-  }
-  
-
-const Template = () => {
-  const [activeCard, setActiveCard] = useState(null);
-  const [prompt, setPrompt] = useState(null);
-  const [isPromptOpen, setIsPromptOpen] = useState(false);
-  const [searchQuery, setSearchQuery] = useState('');
-  const [category, setCategory] = useState('All');
-  const [promptInput, setPromptInput] = useState("");
-  
-  const cards = [
-    { name: "PlatformerGPT  🎮", category: "Games", description: "Write some code to make a platformer game.", promptTemplate: "Write some code to make a platformer game about 'Mario'.", placeholder: "Mario"},
-    { name: "ResearchGPT  📜", category: "Academics", description: "Create a comprehensive report of a topic of your choice", promptTemplate: "Create a comprehensive report of 'Global Warming'.", placeholder: "Global Warming" },
-    { name: "TravelGPT  🌴", category: "Travel", description: "Plan a detailed trip to destination", promptTemplate: "Plan a detailed trip to 'Paris'.", placeholder: "Paris" },
-    { name: "ScraperGPT  ⚒️", category: "Web", description: "Scrap a website of your choice", promptTemplate: "Scrap 'IMDb' website and summarize the details", placeholder: "IMDb" },
-    { name: "PostGPT  📮", category: "Social Media", description: "Thinks of captions and hashtags for your social media posts", promptTemplate: "Create a caption and hashtags for a social media post about 'Summer Vacation'.", placeholder: "Summer Vacation" },
-    { name: "EmailGPT  📧 ", category: "Communications", description: "Compose a concise and detailed email", promptTemplate: "Compose a concise email about 'Project Update'.", placeholder: "Project Update" },
-    { name: "ResumeGPT  📝", category: "Career", description: "Create a professional resume based on your career history.", promptTemplate: "Create a resume detailing your experience in 'Marketing'.", placeholder: "Marketing" },
-    { name: "NovelGPT  📚", category: "Creative Writing", description: "Start writing a novel in your chosen genre.", promptTemplate: "Start writing a 'Fantasy' genre novel", placeholder: "Fantasy" },
-    { name: "DietGPT  🥗", category: "Health & Wellness", description: "Plan a personalized diet based on your dietary preferences.", promptTemplate: "Plan a 'Vegetarian' diet", placeholder: "Vegetarian" },
-    { name: "FitnessGPT  🏋️", category: "Health & Wellness", description: "Design a workout regimen based on your fitness goals.", promptTemplate: "Design a workout regimen for 'Weight Loss'.", placeholder: "Weight Loss" },
-    { name: "MarketingGPT  📈", category: "Business", description: "Create a comprehensive marketing plan for your business.", promptTemplate: "Create a comprehensive marketing plan for 'Startup'.", placeholder: "Startup" },
-    { name: "BudgetGPT  💰", category: "Personal Finance", description: "Prepare a personal or family budget plan.", promptTemplate: "Prepare a budget for 'Family Vacation'.", placeholder: "Family Vacation" },
-    { name: "StudyGPT  📖", category: "Academics", description: "Create a study schedule based on your academic goals.", promptTemplate: "Create a study schedule for 'Final Exams'.", placeholder: "Final Exams" },
-    { name: "NewsGPT  📰", category: "News & Media", description: "Write a detailed news article on a topic of your choice.", promptTemplate: "Write a news article on 'Technology Advancements'.", placeholder: "Technology Advancements" },
-    { name: "EventPlannerGPT  🎉", category: "Event Planning", description: "Plan a detailed schedule for your upcoming event.", promptTemplate: "Plan a detailed schedule for 'Music Festival' event", placeholder: "Music Festival" },
-    { name: "BlogGPT  📝", category: "Blogging", description: "Compose a blog post on a topic of your choice.", promptTemplate: "Compose a blog post about 'Healthy Living'.", placeholder: "Healthy Living" }
-  ];
-  
-
-const filteredCards = cards.filter(card => 
-    (card.name.toLowerCase().includes(searchQuery.toLowerCase()) ||
-    card.description.toLowerCase().includes(searchQuery.toLowerCase()) ||
-    card.category.toLowerCase().includes(searchQuery.toLowerCase())) &&
-    (category === 'All' || card.category === category)
-  );
-
+const Templates = () => {
+ 
 const deployPrompt = () => {
   const currentCard = cards[activeCard];
   if (currentCard) {
@@ -68,112 +20,27 @@ const deployPrompt = () => {
     setIsPromptOpen(false);
   }
 };
-
-  
-
   return (
     <SidebarLayout>
-      <AppTitle/>
-      <div className="container mx-auto px-4 sm:px-6 lg:px-8">
-        <div className="m-4 bg-black shadow-md rounded-lg p-2 flex flex-col sm:flex-row items-start sm:items-center space-y-2 sm:space-y-0 sm:space-x-2">
-            <div className="flex-grow flex space-x-2 w-full">
-                <input
-                type="search"
-                className="flex-grow px-3 py-1 border border-gray-300 rounded-md placeholder-gray-500 shadow-sm focus:outline-none focus:ring-indigo-500 focus:border-indigo-500"
-                placeholder="Search"
-                aria-label="Search"
-                aria-describedby="button-addon2"
-                onChange={(e) => setSearchQuery(e.target.value)}
-                />
-                <button className="hidden sm:inline-flex items-center px-3 py-1 border border-transparent text-sm font-medium rounded-md text-white bg-indigo-600 hover:bg-indigo-700 focus:outline-none focus:ring-2 focus:ring-offset-2 focus:ring-indigo-500">
-                Enter
-                </button>
-            </div>
-            <div className="w-full sm:w-auto">
-                <select id="category" name="category" className="mt-1 block w-full py-1 px-2 border border-gray-300 bg-white rounded-md shadow-sm focus:outline-none focus:ring-indigo-500 focus:border-indigo-500 sm:text-sm" onChange={(e) => setCategory(e.target.value)}>
-                    <option>All</option>
-                    <option>Games</option>
-                    <option>Academics</option>
-                    <option>Travel</option>
-                    <option>Web</option>
-                    <option>Social Media</option>
-                    <option>Communications</option>
-                    <option>Career</option>
-                    <option>Creative Writing</option>
-                    <option>Health & Wellness</option>
-                    <option>Business</option>
-                    <option>Personal Finance</option>
-                    <option>News & Media</option>
-                    <option>Event Planning</option>
-                    <option>Blogging</option>
-                </select>
-            </div>
-        </div>
-
-        <div className="flex items-center justify-center min-h-[200px] ">
-          <div className="grid grid-cols-1 sm:grid-cols-2 md:grid-cols-3 gap-4">
-          {filteredCards.map((card, index) => (
-            <Expand
-                className={`p-2 aspect-content aspect-w-1 min-h-[125px]  border-white rounded bg-${activeCard === index ? 'black' : 'black'} transform transition-transform duration-500 ease-in-out focus:outline-none focus:ring-2 focus:ring-blue-500 cursor-pointer shadow-2xl`}
-                key={index}
-                onClick={() => {
-                setActiveCard(index);
-                setIsPromptOpen(true);
-                }}
-            >
-                    <div className={`text-xs mb-1 inline-block px-2 ${categoryColors[card.category]} rounded-full`}>{card.category}</div>
-                    <div className={`h-full overflow-auto ${activeCard === index ? 'scale-100' : ''}`}>
-                    <div className={`font-bold text-lg text-${activeCard === index ? 'white' : 'white'} mb-2`}>{card.name}</div>
-                    <div className={`text-sm text-${activeCard === index ? 'white' : 'white'}`}>
-                        {activeCard === index ? card.promptTemplate : card.description}
-                    </div>
-                </div>
-            </Expand>
-            ))}
-        </div> 
-      </div>
-    </div>
-    {isPromptOpen && (
-      <div className="fixed inset-0 bg-black bg-opacity-50 flex items-center justify-center z-50 transition-all duration-300 ease-in-out">
-        <div className="p-4 border w-80 h-70 border-gray-200 rounded shadow-lg transform scale-110 bg-black cursor-pointer transition-all duration-300 ease-in-out">
-          <div className="space-y-4">
-            <div className="font-bold text-lg text-white">{cards[activeCard]?.name}</div>
-            <div className={`text-xs mb-1 inline-block px-2 ${categoryColors[cards[activeCard]?.category]} rounded-full`}>{cards[activeCard]?.category}</div>
-            <div className="text-sm text-white">{cards[activeCard]?.promptTemplate}</div>
-            <input
-                type="search"
-                className="w-full px-3 py-2 text-black placeholder-gray-500 border border-gray-300 rounded-md focus:outline-none focus:ring-2 focus:ring-blue-600 focus:border-transparent"
-                placeholder="Enter your variable"
-                aria-label="Variable"
-                aria-describedby="button-addon2"
-                onChange={(e) => setPromptInput(e.target.value)}
-            />
-
-
-
-            <div className="flex justify-end space-x-2">
-              <button
-                className="px-4 py-2 text-white bg-red-500 rounded hover:bg-red-600 focus:outline-none focus:ring-2 focus:ring-red-500"
-                onClick={() => setIsPromptOpen(false)}
-              >
-                Cancel
-              </button>
-              <button
-                className="px-4 py-2 text-white bg-green-500 rounded hover:bg-green-600 focus:outline-none focus:ring-2 focus:ring-green-500"
-                onClick={deployPrompt}
-              >
-                Deploy
-              </button>
-            </div>
+      <div className="w-full h-full p-10">
+        <FadeIn initialX={-45} initialY={0}>
+          <h1 className="text-4xl text-white font-bold">Templates</h1>
+          <h2 className="text-xl text-white font-thin mb-4">Customizable and ready to deploy agents</h2>
+        </FadeIn>
+        <FadeIn initialY={45}>
+          
+          <div className="grid grid-cols-1 sm:grid-cols-2 lg:grid-cols-3 xl:grid-cols-4 gap-4">
+              {TEMPLATE_DATA.map(model => (
+                <TemplateCard key={model.name + model.value} model={model}/>
+                ))}
           </div>
-        </div>
-      </div>
-    )}
+        </FadeIn>
+    </div>
   </SidebarLayout>
   );
 };
 
-export default Template;
+export default Templates;
 
 export const getStaticProps: GetStaticProps = async ({ locale = "en" }) => {
   const supportedLocales = languages.map((language) => language.code);
